Refactor search tools module by removing `agentic_search.py` and updating imports in `searxng_search.py` and `semantic_scholar_search.py` to use the new `_base.py` structure. This change streamlines the code organization and enhances maintainability.
